pyproject.toml
[tool.poetry]
name = "forecast weather site"
version = "0.1.0"
description = "This is the weather site, builded on the Flask"
authors = ["pentagon846 <pentagon846@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
